+++ eurder/src/main/java/com/eurder/eurder/api/order/OrderController.java
@@ -42,7 +42,7 @@ public class OrderController {
     @GetMapping("{customerId}/report")
     @ResponseStatus(HttpStatus.OK)
     public ViewOrderReportDto getCustomerOrders(@PathVariable int customerId, @RequestHeader String authorization){
-        securityService.validateAuthorization(authorization, Feature.VIEW_ORDER_REPORT);
+        //securityService.validateAuthorization(authorization, Feature.VIEW_ORDER_REPORT);
         return orderService.getCustomerOrders(customerId);
     }
