+++ eurder/src/main/resources/application.properties
@@ -2,6 +2,7 @@ spring.datasource.url=jdbc:postgresql://database-2.cuvv0osxzgmi.eu-west-3.rds.am
 spring.datasource.username=student
 spring.datasource.password=spidermanismyhero
 
+spring.jackson.constructor-detector=use_properties_based
 spring.datasource.hikari.minimumIdle=1
 spring.datasource.hikari.maximumPoolSize=3
