+++ eurder/src/main/resources/db/migration/V1__database_setup_create_tables.sql.sql
@@ -1,12 +1,6 @@
-create sequence customer_id_seq
-    as integer
-    maxvalue 9999999;
-
-alter sequence customer_id_seq owner to student;
-
 create table customer
 (
-    id        integer default nextval('"eurder".customer_id_seq'::regclass) not null
+    id        integer generated always as identity
         primary key,
     firstname varchar(255)                                                 not null,
     lastname  varchar(255)                                                 not null,
@@ -15,11 +9,6 @@ create table customer
     address   varchar(255)                                                 not null,
     phone     varchar(255)                                                 not null
 );
-
-alter table customer
-    owner to student;
-
-alter sequence customer_id_seq owned by customer.id;
 
 create table item
 (
@@ -32,9 +21,6 @@ create table item
     urgency_indicator varchar(255)     not null
 );
 
-alter table item
-    owner to student;
-
 create table "order"
 (
     id          integer generated always as identity
@@ -42,9 +28,6 @@ create table "order"
     customer_id integer not null,
     order_date  date    not null
 );
-
-alter table "order"
-    owner to student;
 
 create table item_group
 (
@@ -63,6 +46,4 @@ create table item_group
     item_price          double precision not null
 );
 
-alter table item_group
-    owner to student;
