+++ eurder/src/test/resources/application.properties
@@ -1,0 +1,12 @@
+spring.jpa.hibernate.ddl-auto=none
+spring.flyway.enabled=false
+
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+
+spring.jackson.constructor-detector=use_properties_based
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
